% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_deff.R
\name{calculate_deff}
\alias{calculate_deff}
\title{Calculate design effect}
\usage{
calculate_deff(weight, include_zeroes = FALSE)
}
\arguments{
\item{weight}{A vector of weights. Non-numeric weights will be coerced to numeric.}

\item{include_zeroes}{Should zero weights be included when calculating \code{n},
\code{sd_wt}, and \code{deff}? Defaults to \code{FALSE}.}
}
\value{
A tibble containing the following columns:

\code{n} The sample size

\code{sd_wt} The standard deviation of the weights

\code{deff} The Kish approximation of an overall survey design effect

\code{ess} The effective sample size

\code{moe} The design-adjusted margin of error at 95% confidence
}
\description{
Takes a vector of weights and calculates the design effect, or the proportional increase
in variance due to weighting. Calculated using the Kish (1965) approximation.
}
\examples{
calculate_deff(dec13_excerpt$weight)

}
