% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dk_to_na.R
\name{dk_to_na}
\alias{dk_to_na}
\title{DK to NA}
\usage{
dk_to_na(
  f,
 
    pattern = "^[Dd]on['’]t [Kk]|^[Dd][Kk]|^[Rr]efuse|^[Dd][Kk]/[Rr]ef|\\\\(VOL\\\\.\\\\)$"
)
}
\arguments{
\item{f}{A factor variable.}

\item{pattern}{A regular expression that matches factor levels to convert to NA.}
}
\value{
The factor variable, with DK/Ref values set to NA.
}
\description{
Takes a factor and converts "Don't know/Refused" responses to \code{NA}.
Matches a variety of common patterns such as "Don't know", "Refused"s, "DK/Ref", etc.
}
\details{
Make sure to tabulate the variable after running this function and check
that levels weren't unintentionally removed.
}
\examples{
demonstration <- factor(c(rep("Yes", 57), rep("No", 45), rep("Don't know", 4),
                        rep("Refused", 2), rep("DK/Ref", 10), rep("Not sure (VOL.)", 5)))
tablena(demonstration)
demonstration <- dk_to_na(demonstration)
tablena(demonstration) # success!

}
