% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummify_factors.R
\name{dummify_factors}
\alias{dummify_factors}
\title{Convert all categorical variables into dummy/indicator variables}
\usage{
dummify_factors(df, dummify_characters = TRUE, max_levels = 52, sep = "__")
}
\arguments{
\item{df}{A data frame containing the factors to be converted to dummies.}

\item{dummify_characters}{Should \code{character} variables be converted to factors and
dummified? Defaults to \code{TRUE}.}

\item{max_levels}{The maximum number of levels that a categorical variable can have in
order to be converted to dummies. This is to prevent converting variables such as
respondent IDs or open-ended responses to dummy variables. Defaults to 52, which is
slightly more than the number of U.S. states plus Washington, D.C.}

\item{sep}{A character string that will go in between the original variable name and the
corresponding factor level in the output, e.g. VARIABLE__Category. Defaults to \code{"__"}.}
}
\value{
A copy of the original data frame where all factor variables have been replaced
with dummies. Formatted as a tibble.
}
\description{
Takes a data frame and separates all columns containing categorical variables into
dummy variables (one column per category). Unlike model.matrix(), does not exclude any
categories from being converted into columns, and retains missing values.
}
\examples{
test <- dummify_factors(dec13_excerpt)
head(test)

}
