% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spss_label.R
\name{get_spss_label}
\alias{get_spss_label}
\title{Get SPSS label}
\usage{
get_spss_label(df, var, null_label = "No label found", unlist = TRUE)
}
\arguments{
\item{df}{An SPSS dataset originally read into R via either \code{foreign} or
\code{haven}.}

\item{var}{\code{character} vector containing the names of variables whose labels
should be returned.}

\item{null_label}{\code{character} value to return if there is no label for a variable.
Defaults to \code{"No label found"}.}

\item{unlist}{Should the output be converted from a list to a vector? Defaults to
\code{TRUE}. Set to \code{FALSE} to return a list instead.}
}
\value{
The SPSS label/s, as a vector or a list.
}
\description{
Returns variable labels from an SPSS dataset read in with either the \code{foreign}
or \code{haven} packages.
}
\examples{
get_spss_label(dec13_excerpt, "q1")
get_spss_label(dec13_excerpt, c("q1", "party"))
get_spss_label(dec13_excerpt, c("q1", "sdfklsdf", "party"), null_label = "here be dragons")
get_spss_label(dec13_excerpt, c("q1", "sdfklsdf", "party"), unlist = FALSE)
}
