% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_totals.R
\name{get_totals}
\alias{get_totals}
\title{Get weighted percentages or totals}
\usage{
get_totals(
  var,
  df,
  wt = NULL,
  by = NULL,
  by_total = FALSE,
  percent = TRUE,
  include_unw = FALSE,
  digits = NULL,
  complete = TRUE,
  na.rm = FALSE
)
}
\arguments{
\item{var}{\code{character}, indicating the name of the variable to be tabulated.
Specify interactions by separating variable names with a colon.}

\item{df}{The \code{data.frame} containing the variables to be tabulated.}

\item{wt}{A \code{character} vector containing the name(s) of the weight variable(s)
to be used in tabulating results. If nothing is passed, results will be unweighted.}

\item{by}{For creating crosstabulations, an optional \code{character} variable
containing the name of the variable to be crossed with \code{var}. Can pass
multiple variables in as a vector.}

\item{by_total}{\code{logical} indicating whether a "total" column should be
returned in addition to columns defined by variables passed to \code{by}.
Defaults to \code{FALSE}.}

\item{percent}{Should the results be scaled as percentages? Defaults to \code{TRUE}.
If \code{FALSE}, weighted totals are returned.}

\item{include_unw}{Include unweighted frequencies in the output along with weighted.
Defaults to \code{FALSE}}

\item{digits}{The number of decimal points displayed. Defaults to value specified
in \code{options("digits")}.}

\item{complete}{TRUE/FALSE: Should factor levels with no
observations be included in the results? Defaults to \code{TRUE}.}

\item{na.rm}{If \code{FALSE}, \code{NA} values in \code{var} are included in the
results and included in the denominator for calculating percentages. If \code{TRUE},
they are excluded from any calculations. Defaults to \code{FALSE}.}
}
\value{
A \code{data.frame} with a column for the variable name, columns displaying
percentages or totals, and additional columns as specified by arguments to
this function.
}
\description{
Takes a categorical variable and returns the weighted or unweighted
percentage or total of each category. Can include a grouping variable. Can be
used to compare weights versus one another.
}
\details{
If no arguments are supplied to \code{by}, then the column names
  will be the weight names. If arguments are supplied to \code{by}, then the
  column names will be the categories of the grouping variable, and the
  output will have an additional column for the weight name.
}
\examples{
library(dplyr)
# Basic unweighted crosstab
get_totals("q1", dec13_excerpt)

# Totals instead of percentages
get_totals("q1", dec13_excerpt, percent = FALSE)

# Weighted crosstab
get_totals("receduc", dec13_excerpt, wt = "weight")

# Weighted crosstab by grouping variable
get_totals("q1", dec13_excerpt, wt = "weight", by = "receduc")

# Compare weights, including unweighted
# Let's make a fake weight by combining the landline and cellphone weights
dec13_excerpt <- dec13_excerpt \%>\% mutate(fake_weight = coalesce(llweight, cellweight))
get_totals("q1", dec13_excerpt, wt = c("weight", "fake_weight"), include_unw = TRUE)

# Use dplyr::filter along with complete = FALSE to remove unwanted categories from the base
get_totals("q1", dec13_excerpt \%>\% filter(q1 != "Don't know/Refused (VOL.)"), wt = "weight",
           complete = FALSE)

# Alternatively, filter unwanted categories out beforehand with dk_to_na
# and then use na.rm = TRUE
dec13_excerpt <- dec13_excerpt \%>\% mutate(q1 = dk_to_na(q1))
get_totals("q1", dec13_excerpt, wt = "weight", na.rm = TRUE)

}
