% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboardize_punctuation.R
\name{keyboardize_punctuation}
\alias{keyboardize_punctuation}
\title{Keyboardize punctuation marks}
\usage{
keyboardize_punctuation(df, convert_characters = FALSE)
}
\arguments{
\item{df}{A \code{data.frame}.}

\item{convert_characters}{Should variables of class \code{character} also be converted,
in addition to variables of class "factor"? Defaults to \code{FALSE}.}
}
\value{
A copy of the original \code{data.frame}, with nonstandard characters replaced.
}
\description{
Takes a dataset and, for all categories in factor variables, standardizes stylized
typographic characters to ASCII versions that are easily typeable on a keyboard.
}
\details{
Text labels in datasets occasionally contain characters not easily typeable on a standard
keyboard which are almost impossible to distinguish from keyboard characters. (This often
happens when pasting text in from Microsoft Word). These include the curly apostrophe,
the en-dash, the em-dash and the ellipsis. This function converts these typographic
characters to characters that are on keyboards by default: the single straight apostrophe,
the hyphen-minus, and three periods.
}
\examples{
library(dplyr)
# Create an unwanted factor label containing a stylized apostrophe ’ for illustration.
dec13_excerpt <- dec13_excerpt \%>\% mutate(receduc =
                                           forcats::fct_recode(receduc,
                                                               `Don’t know/Refused` = "DK/Ref"))
table(dec13_excerpt$receduc)
dec13_excerpt <- keyboardize_punctuation(dec13_excerpt)
table(dec13_excerpt$receduc)

}
