% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.ranger.R
\name{mice.impute.ranger}
\alias{mice.impute.ranger}
\title{Random forest imputation using the ranger package}
\usage{
mice.impute.ranger(y, ry, x, ntree = 10, wy = NULL, type = NULL, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} marking observed values of y
with TRUE and missing values of y with FALSE.}

\item{x}{the matrix of predictors, without intercept}

\item{ntree}{the number of trees for the random forest algorithm. Default is 10.}

\item{wy}{Logical vector of length \code{length(y)}. A TRUE value indicates locations in y for which imputations are created.}

\item{type}{Not used. Needed for compatibility with mice}

\item{...}{Other arguments passed to \code{ranger()}}
}
\value{
A vector of imputed values.
}
\description{
Designed as an alternative to the the \code{"rf"} method in the \code{mice} package.
Imputes univariate missing data using random forests. It uses the same algorithm as
\code{"mice.impute.rf"} (included in the \code{mice} package, but is much faster because it uses
the \code{ranger} package.
This function should not be called directly.
Instead, call this function through \code{mice()} by specifying \code{method = "ranger"}.
}
