% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pewmethods.R
\docType{package}
\name{pewmethods}
\alias{pewmethods}
\title{pewmethods: Miscellaneous functions by the Pew Research Center Methods team}
\description{
The \code{pewmethods} package contains functions that have been used internally by the Methods team
at Pew Research Center to make everyday work with survey data easier. These include functions to
quickly summarize weighted data, clean and recode variables, create survey weights (wrapping around the
\code{survey} package), output crosstabs, and a number of other common tasks that arise in the analysis of
survey data.
}
\details{
For an overview of what's in \code{pewmethods}, browse our vignettes by calling \code{vignette("pewmethods_exploring")}
and \code{vignette("pewmethods_weighting")}.
}
\section{Author(s)}{

\itemize{
\item \strong{Maintainer:} Arnold Lau \href{mailto:alau@pewresearch.org}{alau@pewresearch.org}
\item Andrew Mercer \href{mailto:amercer@pewresearch.org}{amercer@pewresearch.org}
\item Nick Hatley \href{mailto:nhatley@pewresearch.org}{nhatley@pewresearch.org}
}
}

