% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rake_survey.R
\name{rake_survey}
\alias{rake_survey}
\title{Rake survey}
\usage{
rake_survey(
  .data,
  pop_margins,
  base_weight = 1,
  scale_to_n = TRUE,
  epsilon = 5e-06,
  maxit = 100,
  ...
)
}
\arguments{
\item{.data}{A \code{data.frame} containing the survey data}

\item{pop_margins}{A list of tibbles giving the population margins for raking variables.
\code{create_raking_targets} outputs such a list in the correct format.}

\item{base_weight}{The survey's base weight variable, if applicable. Can be ignored if the survey
data doesn't come with a base weight.}

\item{scale_to_n}{If \code{TRUE}, scales the resulting weights so so that they sum to the number of
observations in the survey.}

\item{epsilon}{A \code{numeric} variable indicating the tolerance for the raking algorithm. When the
proportions of the raking variables are within tolerance, the raking algorithm will stop.}

\item{maxit}{Maximum number of iterations before stopping if raking has not converged.}

\item{...}{Other arguments passed to \code{survey::calibrate()}.}
}
\value{
A vector of survey weights.
}
\description{
Applies the raking algorithm to a survey using specified raking targets in order to obtain weights.
Wrapper around the \code{calibrate} function from the \code{survey} package with argument
\code{calfun = "raking"}.
}
\details{
The variables in \code{.data} must exactly match the variables in \code{pop_margins} both
in name in the factor values. Additionally, they should contain no missing data.
}
\examples{
library(dplyr)
# Prepare variables from the survey data for raking
dec13_excerpt_raking <- dec13_excerpt \%>\%
  mutate(rk_sex = sex,
         rk_recage = dk_to_na(recage, pattern = "DK/Ref"),
         rk_receduc = dk_to_na(receduc, pattern = "DK/Ref")) \%>\%
  impute_vars(.) \%>\%
  mutate(rk_sex_receduc = interaction(rk_sex, rk_receduc, sep = ":"))
# Prepare population marginal distributions for raking
# Here we will use the acs_2017_excerpt dataset included wih the package
targets <- create_raking_targets(acs_2017_excerpt,
                                 vars = c("sex", "recage", "receduc", "sex:receduc"),
                                 wt = "weight")

# Now that we have raking variables and population targets, we can create a raking weight
fake_weight <- rake_survey(dec13_excerpt_raking, targets)

}
