% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_if.R
\name{replace_if}
\alias{replace_if}
\title{Replace if}
\usage{
replace_if(var, condition, replacement)
}
\arguments{
\item{var}{A variable.}

\item{condition}{Either a scalar, a logical vector of the same length as \code{var},
or NA (in which case the function will look for and replace all NA values).}

\item{replacement}{A scalar.}
}
\value{
\code{var}, replaced.
}
\description{
Takes a vector and replaces elements in the vector according to some condition.
(Equivalent to something like \code{df$var1[df$var2 == TRUE] <- 4} but less clunky.)
Unlike the base function \code{replace}, \code{replace_if} allows replacement of NAs
 with some other value and only accepts scalar values as replacements (preventing
 recycling value errors). As with \code{replace}, this function does not work with
 factors; use \code{forcats::fct_recode}.
}
\examples{
x <- c(1, 2, 3, 4)
replace_if(var = x, condition = x > 2, replacement = 99)

x <- c(1, 2, NA, 4)
replace_if(var = x, condition = NA, replacement = 99)

}
