% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_spss_label.R
\name{set_spss_label}
\alias{set_spss_label}
\title{Set SPSS label}
\usage{
set_spss_label(df, var, label_text)
}
\arguments{
\item{df}{An SPSS dataset read into R via either \code{foreign} or \code{haven}.}

\item{var}{\code{character} vector containing the names of variables to be labeled.}

\item{label_text}{\code{character} vector containing the label or labels to be assigned.}
}
\value{
A copy of \code{df} with variable labels updated.
}
\description{
Writes a label to an SPSS dataset read into R with either the \code{foreign} or \code{haven}
packages.
}
\examples{
dec13_excerpt <- set_spss_label(dec13_excerpt, "q1", "Obama approval")
dec13_excerpt <- set_spss_label(dec13_excerpt,
                               c("q1", "q45"),
                               c("Obama approval", "Obamacare approval"))

# Let's create a fake new variable for illustration
dec13_excerpt$newvar <- 1
dec13_excerpt <- set_spss_label(dec13_excerpt, "newvar", "New variable")
}
