% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_weights.R
\name{standardize_weights}
\alias{standardize_weights}
\title{Standardize weights}
\usage{
standardize_weights(weight, to_ess = FALSE, na_to_zero = TRUE)
}
\arguments{
\item{weight}{A vector of weights. Coerced to \code{numeric} if a different type.}

\item{to_ess}{Should the weights sum to the effective sample size? Defaults to \code{FALSE}.}

\item{na_to_zero}{Should missing weights become zero weights? Defaults to \code{TRUE}.}
}
\value{
A vector of standardized weights.
}
\description{
Divides each weight by the mean and converts missing weights to 0.
The resulting weights will have a mean of 1 and sum to either the nominal sample size
or the effective sample size, which is the nominal sample size divided by the Kish (1965)
approximation of an overall survey design effect
}
\examples{
summary(dec13_excerpt$weight)
dec13_excerpt$weight <- standardize_weights(dec13_excerpt$weight)
summary(dec13_excerpt$weight)

}
