% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syssamp.R
\name{syssamp}
\alias{syssamp}
\title{Systematic sample from a dataset}
\usage{
syssamp(.data, size, stratvars, return_indices = TRUE, seed = NA)
}
\arguments{
\item{.data}{A dataset, in data.frame or tibble format.}

\item{size}{The size of the desired sample.}

\item{stratvars}{Names of variables by which the sampling frame will be sorted}

\item{return_indices}{Return a vector containing the indices of the selected records. Defaults to \code{TRUE}.
If \code{FALSE}, a data.frame containing the sampled records will be returned.}

\item{seed}{Optional. Ensures that the same sample will be created if the code is rerun.}
}
\value{
Either a vector of indices or a data.frame/tibble.
}
\description{
Takes a dataset, sorts it by one or more variables, and draws samples at regular intervals.
A fractional skip interval is used so that each record still has an equal probability of selection
when the number of records in the frame is not evenly divisible by the sample size.
}
\examples{
dec13_sample_indices <- syssamp(dec13_excerpt, size = 10, stratvars = c("sex", "recage"))
dec13_sample <- syssamp(dec13_excerpt, size = 10, stratvars = c("sex", "recage"),
                        return_indices = FALSE)

}
\references{
Buskirk, T.D. (2008) Sampling interval. In P.J. Lavrakas, \emph{Encyclopedia of survey research methods}. Thousand Oaks, CA: Sage Publications.
}
