% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablena.R
\name{tablena}
\alias{tablena}
\title{Enhanced tables}
\usage{
tablena(..., show_class = TRUE)
}
\arguments{
\item{...}{Arguments passed to \code{table()}}

\item{show_class}{TRUE by default. Shows the class of the variable being tabled.}
}
\value{
A table
}
\description{
Displays tables with NAs, dimension names, and dimension classes by default
}
\examples{
tablena(dec13_excerpt$q2, dec13_excerpt$receduc)
# Also works with the with() function and other such functions that construct environments from data
with(dec13_excerpt, tablena(q2, receduc))

}
