% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timefactory.R
\name{timefactory}
\alias{timefactory}
\title{Create running timer functions}
\usage{
timefactory()
}
\value{
A function that will return the number of seconds since it
was first created.
}
\description{
This is a factory function for creating running timers. A function
created using \code{timefactory()} will return the number of seconds
since it was first created. It is intended as a simple alternative to
wrapping code in multiple calls to \code{proc.time()}.
}
\examples{
# Create an initial timer
timer1 = timefactory()
Sys.sleep(3)
timer1()

# Create a second timer that is independent of the first.
timer2 = timefactory()
Sys.sleep(2)

timer1()
timer2()

}
