% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_weights.R
\name{trim_weights}
\alias{trim_weights}
\title{Trim weights}
\usage{
trim_weights(
  weight,
  lower_quantile = 0.01,
  upper_quantile = 0.99,
  minval = NULL,
  maxval = NULL,
  strict = FALSE
)
}
\arguments{
\item{weight}{A numeric vector of weights}

\item{lower_quantile}{The quantile at which the weights should be trimmed on the lower end}

\item{upper_quantile}{The quantile at which the weights should be trimmed on the upper end}

\item{minval}{NULL by default. If the lower quantile is less than the minimum value, minval will be used
for the trim.}

\item{maxval}{NULL by default. If the upper quantile is greater than the maximum value, maxval will be used
for the trim.}

\item{strict}{When the trimmed weights are reapportioned, some values may exceed the specificed threshold value.
If \code{strict = FALSE} this is permitted. Otherwise, weights are trimmed recursively until none exceed the limit.}
}
\value{
A numeric vector of trimmed weights
}
\description{
Trim survey weights. Wrapper for the \code{trimWeights()} function in the \code{survey} package that eliminates
the need to create a \code{svydesign} object.
}
\examples{
calculate_deff(dec13_excerpt$weight)
trimmed <- trim_weights(dec13_excerpt$weight, lower_quantile = 0.1, upper_quantile = 0.9)
# Trimming reduces the standard deviation of the weights, reducing reducing the design effect
# and increasing the effective sample size.
calculate_deff(trimmed)

}
