% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_margins.R
\name{unify_margins}
\alias{unify_margins}
\title{Unify margins}
\usage{
unify_margins(pop_margins, .data)
}
\arguments{
\item{pop_margins}{A list of tibbles giving the population margins for raking variables. \code{create_raking_targets()} outputs such a list.}

\item{.data}{The survey data to be raked.}
}
\value{
Cleaned raking targets.
}
\description{
This is an internal function called by \code{rake_survey()} that makes sure that the raking targets and the survey data match up.
}
\details{
Ensures that the factor levels for each variable in the targets and in the data are in the same order,
and that the list of targets is not named.
}
